backend/text_extractor_api/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
PyMuPDF
python-docx